add utils library for calc_shifts_required function
