+++ libs/utils/aframe/utils/timeslides.py
@@ -1,0 +1,25 @@
+import math
+
+
+def calc_shifts_required(Tb: float, T: float, delta: float) -> int:
+    r"""
+    Calculate the number of shifts required to generate Tb
+    seconds of background.
+
+    The algebra to get this is gross but straightforward.
+    Just solving
+    $$\sum_{i=1}^{N}(T - i\delta) \geq T_b$$
+    for the lowest value of N, where \delta is the
+    shift increment.
+
+    TODO: generalize to multiple ifos and negative
+    shifts, since e.g. you can in theory get the same
+    amount of Tb with fewer shifts if for each shift
+    you do its positive and negative. This should just
+    amount to adding a factor of 2 * number of ifo
+    combinations in front of the sum above.
+    """
+
+    discriminant = (delta / 2 - T) ** 2 - 2 * delta * Tb
+    N = (T - delta / 2 - discriminant**0.5) / delta
+    return math.ceil(N)
