+++ libs/utils/pyproject.toml
@@ -1,0 +1,16 @@
+[tool.poetry]
+name = "aframe.utils"
+version = "0.1.0"
+description = "aframe utility functions"
+authors = ["Ethan Marx <emarx@mit.edu>"]
+license = "MIT"
+#readme = "README.md"
+packages = [{include = "aframe"}]
+
+[tool.poetry.dependencies]
+python = ">=3.8,<3.11"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
