+++ projects/sandbox/datagen/datagen/utils/timeslide_waveforms.py
@@ -1,4 +1,3 @@
-import math
 import time
 from pathlib import Path
 from typing import List, Tuple
@@ -7,29 +6,7 @@ import numpy as np
 import torch
 from gwpy.timeseries import TimeSeriesDict
 
-
-def calc_shifts_required(Tb: float, T: float, delta: float) -> int:
-    r"""
-    Calculate the number of shifts required to generate Tb
-    seconds of background.
-
-    The algebra to get this is gross but straightforward.
-    Just solving
-    $$\sum_{i=1}^{N}(T - i\delta) \geq T_b$$
-    for the lowest value of N, where \delta is the
-    shift increment.
-
-    TODO: generalize to multiple ifos and negative
-    shifts, since e.g. you can in theory get the same
-    amount of Tb with fewer shifts if for each shift
-    you do its positive and negative. This should just
-    amount to adding a factor of 2 * number of ifo
-    combinations in front of the sum above.
-    """
-
-    discriminant = (delta / 2 - T) ** 2 - 2 * delta * Tb
-    N = (T - delta / 2 - discriminant**0.5) / delta
-    return math.ceil(N)
+from aframe.utils.timeslides import calc_shifts_required
 
 
 def get_num_shifts(
