+++ projects/sandbox/datagen/pyproject.toml
@@ -18,17 +18,18 @@ generate-timeslide-waveforms = "datagen.scripts.timeslide_waveforms:main"
 deploy-timeslide-waveforms = "datagen.scripts.timeslide_waveforms:deploy"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.11"
+python = ">=3.9,<3.11"
 
 typeo = {git = "https://github.com/ML4GW/typeo.git", branch = "main"}
 torch = {version = "^1.10", source = "torch"}
 ml4gw = {path = "../../../ml4gw", develop = true}
-mldatafind = {git = "https://github.com/ML4GW/mldatafind.git", branch = "main"}
+mldatafind = {path = "../../../mldatafind/", develop = true}
 
 "aframe.analysis" = {path = "../../../libs/analysis", develop = true}
 "aframe.deploy" = {path = "../../../libs/deploy", develop = true}
 "aframe.logging" = {path = "../../../libs/logging", develop = true}
 "aframe.priors" = {path = "../../../libs/priors", develop = true}
+"aframe.utils" = {path = "../../../libs/utils", develop = true}
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.3"
