+++ projects/sandbox/infer/pyproject.toml
@@ -21,6 +21,7 @@ typeo = {git = "https://github.com/ML4GW/typeo.git", branch = "main"}
 "aframe.analysis" = {path = "../../../libs/analysis", develop = true}
 "aframe.deploy" = {path = "../../../libs/deploy", develop = true}
 "aframe.logging" = {path = "../../../libs/logging", develop = true}
+"aframe.utils" = {path = "../../../libs/utils", develop = true}
 
 # hermes utilities
 "hermes.aeriel" = {path = "../../../hermes/hermes/hermes.aeriel", extras=["serve"], develop = true}
